Constructor of an object of the popim_vacc_activites class

The popim_vacc_activities object is a dataframe that holds information on vaccination activities that are typically meant to be applied to a popim_population object.

popim_vacc_activities(
  region = character(),
  year = integer(),
  age_first = integer(),
  age_last = integer(),
  coverage = double(),
  doses = double(),
  targeting = character()
)

Arguments

region

character vector: specifies the geographic region to which each vaccination activity is to be administered.

year

integer vector: specifies the year in which each vaccination activity takes place.

age_first, age_last

non-negative integer vectors: specify the age range targeted in the vaccination activity. age_first <= age_last for all entries.

coverage

numeric vector, 0 <= coverage <= 1 for all entries. Specifies the proportion of the target population to be immunised.

doses

numeric vector, non-negative. Specifies the number of doses to be used in each vaccination activity.

targeting

character vector, permissible entries are "random", "correlated", "targeted". Defines how vaccine is allocated if there is pre-existing immunity in the population: For "random" targeting individuals are vaccinated irrespective of immunity status, so if prior to the vaccination activity the proportion immune was x, then a proportion x of the vaccine will be administered to already vaccinated individuals and therefore be wasted. For "correlated" targeting vaccine is administered first to those already immune before any susceptible individuals receive vaccine. This option models the case of unequal access to vaccination. For "targeted" targeting, vaccine will be given first to as yet non-immune individuals. This is the most effective use of vaccine. It may be realistic in the case of multi-year campaigns targeting different areas within the geographical region specified.

Value

S3 object of class popim_vacc_activities: a dataframe with columns region, year, age_first, age_last, coverage, doses, targeting.

Details

The input parameters are the columns of the popim_vacc_activities object to be returned, they should be vectors of the same length, or will be recycled as appropriate by data.frame().

Author

Tini Garske

Examples

# setting up an empty class `popim_vacc_activities` object of the
# correct structure:
vacc <- popim_vacc_activities()

# setting up a two specific vaccination activities
vacc <- popim_vacc_activities(region = c("UK", "FRA"),
                            year = c(2010, 2005),
                            age_first = c(0, 0),
                            age_last = c(0, 10),
                            coverage = c(0.8, 0.5),
                            doses = NA,
                            targeting = "random")